requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
torch
